Fix: move syncBookHighlights to plugin.js and adjacent refactoring
