+++ lib/__mocks__/readwise.js
@@ -1,0 +1,7 @@
+// Assuming you're using ES6 modules, adjust the syntax accordingly
+function _readwiseMakeRequest() {
+    console.log("iupi");
+}
+
+// Exporting the mocked functions
+export { _readwiseMakeRequest };
