+++ lib/books.js
@@ -1,103 +1,9 @@
-import { _getLastUpdatedTimeFromNote } from './dates.js';
-import { _noteContent,
-  _insertContent,
-  _replaceContent,
-  _sectionContent,
-  _textToTagName
-} from './amplenote_rw.js';
-import { _sectionsFromMarkdown,
-  _markdownFromSections,
-  _markdownFromHighlights,
-} from './markdown.js';
-import { _groupByValue,
-  _distributeIntoSmallGroups,
-  _trimHighlights
-} from './data_structures.js';
-import { _yearFromDateString,
-  _localeDateFromIsoDate
-} from './dates.js';
+import { _textToTagName } from './amplenote_rw.js';
+import { _localeDateFromIsoDate } from './dates.js';
 /*******************************************************************************************/
 /* Book notes
 /*******************************************************************************************/
 
-/*******************************************************************************************
- * Sync highlights for a book into the note provided. This method does all of the propagation from
- * Readwise Highlight object to list of highlights in a note
- *
- * Returns true if successful
- */
-export async function _syncBookHighlights(app, noteContents, forceReprocess, readwiseFetchBooks, constants, dateFormat, bookNote, readwiseBookID, { readwiseBook = null, throwOnFail = false } = {}) {
-  console.log(`_syncBookHighlights(app, ${bookNote}, ${readwiseBookID})`);
-  let lastUpdatedAt = await _getLastUpdatedTimeFromNote(app, constants, bookNote);
-  if (forceReprocess) {
-    lastUpdatedAt = null;
-  }
-
-  // Import all (new) highlights from the book
-  const noteContent = await _noteContent(noteContents, bookNote) || "";
-
-  if (!noteContent.includes("# Summary")) {
-    await _insertContent(noteContents, bookNote, "# Summary\n");
-  }
-
-  if (!readwiseBook) {
-    let generator = _readwiseFetchBooks(app, constants.readwiseConstants, {bookIdFilter: readwiseBookID});
-    let result = await generator.next();
-    readwiseBook = result.value;
-
-    if (!readwiseBook) {
-      if (throwOnFail) {
-        throw new Error(`Could not fetch book details for book ID ${ readwiseBookID }, you were probably rate-limited by Readwise. Please try again in 30-60 seconds?`);
-      } else {
-        return false;
-      }
-    }
-  }
-
-  const summaryContent = _bookNotePrefaceContentFromReadwiseBook(app, constants.bookConstants, dateFormat, readwiseBook, bookNote.uuid);
-  await _replaceContent(noteContents, bookNote, "Summary", summaryContent);
-
-  let highlightsContent = "";
-  if (!noteContent.includes("# Highlights")) {
-    await _insertContent(noteContents, bookNote, "\n# Highlights\n", { atEnd: true });
-  } else {
-    highlightsContent = _sectionContent(noteContent, "Highlights");
-  }
-  let highlights = await _sectionsFromMarkdown(noteContent, "Highlights", _loadHighlights);
-
-  let bookNoteHighlightList = await _bookHighlightsContentFromReadwiseBook(app, readwiseBook, highlights, lastUpdatedAt);
-  const sortOrder = app.settings[constants.bookConstants.settingSortOrderName] || constants.bookConstants.defaultHighlightSort;
-  let hlGroups = _groupByValue(bookNoteHighlightList,
-    item => {
-      if (!item.highlighted_at) return "No higlight date";
-      let year = _yearFromDateString(item.highlighted_at);
-      if (!year) return "No highlight date";
-      return year;
-    },
-  );
-  hlGroups = _distributeIntoSmallGroups(hlGroups, constants.bookConstants.maxBookHighlightsPerSection,
-      constants.bookConstants.maxReplaceContentLength, _trimHighlights, dateFormat);
-  let entries = Object.entries(hlGroups);
-  entries.sort((a, b) => b[0].localeCompare(a[0]));
-  let hlMarkdown = _markdownFromSections(app, entries, _markdownFromHighlights.bind(this));
-
-  try {
-    await _replaceContent(noteContents, bookNote, "Highlights", hlMarkdown);
-  } catch (error) {
-    console.log("Error replacing", readwiseBook.title, "content, length", hlMarkdown.length ," error", error);
-  }
-
-  let existingContent = "";
-  if (!noteContent.includes("Sync History")) {
-    await _insertContent(noteContents, bookNote, "\n# Sync History\n", { atEnd: true });
-  } else {
-    const match = noteContent.match(/#\sSync\sHistory\n([\s\S]+)$/m);
-    existingContent = match ? match[1] : "";
-  }
-
-  await _replaceContent(noteContents, bookNote, "Sync History",`${ constants.bookConstants.updateStringPreface }${ _localeDateFromIsoDate(new Date(), dateFormat) }\n` + existingContent);
-  return true;
-}
 
 /*******************************************************************************************
  * Returns the note handle of a highlight note given a Readwise Book object
