+++ lib/plugin.js
@@ -13,15 +13,21 @@ import { _noteContent,
   _replaceContent,
   _flushLocalNotes,
 } from './amplenote_rw.js';
-import { _sectionsFromMarkdown,
+import {
+  _markdownFromHighlights,
+  _markdownFromSections, _sectionsFromMarkdown,
   _tableFromMarkdown,
 } from './markdown.js';
-import { _groupByValue } from './data_structures.js';
+import {_distributeIntoSmallGroups, _groupByValue, _trimHighlights} from './data_structures.js';
 import * as readwise from './readwise.js';
-import { _ensureBookNote,
+import {
+  _ensureBookNote,
   _bookObjectFromReadwiseBook,
-  _syncBookHighlights,
+  _bookNotePrefaceContentFromReadwiseBook,
+  _loadHighlights,
+  _bookHighlightsContentFromReadwiseBook,
 } from './books.js';
+import {_getLastUpdatedTimeFromNote, _localeDateFromIsoDate, _yearFromDateString} from "./dates.js";
 
 const plugin = {
   // TODO: handle abort execution
@@ -239,7 +245,7 @@ const plugin = {
             await _flushLocalNotes(app, this._noteContents);
           }
         }
-        const success = await _syncBookHighlights(app, this._noteContents, this._forceReprocess, this.readwiseModule._readwiseFetchBooks, this.constants, dateFormat, bookNote, readwiseBook.id, { readwiseBook });
+        const success = await this._syncBookHighlights(app, bookNote, readwiseBook.id, { readwiseBook });
         if (success) bookCount += 1;
       }
       // Edits dashboard note
@@ -286,8 +292,7 @@ const plugin = {
 
       // Import all (new) highlights from the book
       const bookId = match[1];
-      const dateFormat = this._dateFormat || (app && app.settings[this.constants.settingDateFormat]) || "en-US";
-      const success = await _syncBookHighlights(app, this._noteContents, this._forceReprocess, this.readwiseModule._readwiseFetchBooks, this.constants, dateFormat, currentNote, bookId, { throwOnFail: true });
+      const success = await this._syncBookHighlights(app, currentNote, bookId, { throwOnFail: true });
       if (this._useLocalNoteContents) {
         await _flushLocalNotes(app);
       }
@@ -296,7 +301,88 @@ const plugin = {
       }
     } catch (error) {
       await app.alert(String(error));
+      throw(error);
     }
+  },
+
+  /*******************************************************************************************
+   * Sync highlights for a book into the note provided. This method does all of the propagation from
+   * Readwise Highlight object to list of highlights in a note
+   *
+   * Returns true if successful
+   */
+  async _syncBookHighlights(app, bookNote, readwiseBookID, { readwiseBook = null, throwOnFail = false } = {}) {
+    console.log(`_syncBookHighlights(app, ${bookNote}, ${readwiseBookID})`);
+    const dateFormat = this._dateFormat || (app && app.settings[this.constants.settingDateFormat]) || "en-US";
+    let lastUpdatedAt = await _getLastUpdatedTimeFromNote(app, this.constants, bookNote);
+    if (this._forceReprocess) {
+      lastUpdatedAt = null;
+    }
+
+    // Import all (new) highlights from the book
+    const noteContent = await _noteContent(this._noteContents, bookNote) || "";
+
+    if (!noteContent.includes("# Summary")) {
+      await _insertContent(this._noteContents, bookNote, "# Summary\n");
+    }
+
+    if (!readwiseBook) {
+      let generator = this.readwiseModule._readwiseFetchBooks(app, this.constants.readwiseConstants, {bookIdFilter: readwiseBookID});
+      let result = await generator.next();
+      readwiseBook = result.value;
+
+      if (!readwiseBook) {
+        if (throwOnFail) {
+          throw new Error(`Could not fetch book details for book ID ${ readwiseBookID }, you were probably rate-limited by Readwise. Please try again in 30-60 seconds?`);
+        } else {
+          return false;
+        }
+      }
+    }
+
+    const summaryContent = _bookNotePrefaceContentFromReadwiseBook(app, this.constants.bookConstants, dateFormat, readwiseBook, bookNote.uuid);
+    await _replaceContent(this._noteContents, bookNote, "Summary", summaryContent);
+
+    let highlightsContent = "";
+    if (!noteContent.includes("# Highlights")) {
+      await _insertContent(this._noteContents, bookNote, "\n# Highlights\n", { atEnd: true });
+    } else {
+      highlightsContent = _sectionContent(noteContent, "Highlights");
+    }
+    let highlights = await _sectionsFromMarkdown(noteContent, "Highlights", _loadHighlights);
+
+    let bookNoteHighlightList = await _bookHighlightsContentFromReadwiseBook(app, readwiseBook, highlights, lastUpdatedAt);
+    // const sortOrder = app.settings[constants.bookConstants.settingSortOrderName] || constants.bookConstants.defaultHighlightSort;
+    let hlGroups = _groupByValue(bookNoteHighlightList,
+        item => {
+          if (!item.highlighted_at) return "No higlight date";
+          let year = _yearFromDateString(item.highlighted_at);
+          if (!year) return "No highlight date";
+          return year;
+        },
+    );
+    hlGroups = _distributeIntoSmallGroups(hlGroups, this.constants.bookConstants.maxBookHighlightsPerSection,
+        this.constants.bookConstants.maxReplaceContentLength, _trimHighlights, dateFormat);
+    let entries = Object.entries(hlGroups);
+    entries.sort((a, b) => b[0].localeCompare(a[0]));
+    let hlMarkdown = _markdownFromSections(app, entries, _markdownFromHighlights.bind(this));
+
+    try {
+      await _replaceContent(this._noteContents, bookNote, "Highlights", hlMarkdown);
+    } catch (error) {
+      console.log("Error replacing", readwiseBook.title, "content, length", hlMarkdown.length ," error", error);
+    }
+
+    let existingContent = "";
+    if (!noteContent.includes("Sync History")) {
+      await _insertContent(this._noteContents, bookNote, "\n# Sync History\n", { atEnd: true });
+    } else {
+      const match = noteContent.match(/#\sSync\sHistory\n([\s\S]+)$/m);
+      existingContent = match ? match[1] : "";
+    }
+
+    await _replaceContent(this._noteContents, bookNote, "Sync History",`${ this.constants.bookConstants.updateStringPreface }${ _localeDateFromIsoDate(new Date(), dateFormat) }\n` + existingContent);
+    return true;
   },
 
   /*******************************************************************************************/
@@ -309,7 +395,7 @@ const plugin = {
           inputs: [{
             type: "select",
             label: "Category",
-            options: categories.map(function(value, index) {
+            options: categories.map(function(value) {
               return { value: value, label: value };
             })
           }]
@@ -328,12 +414,10 @@ const plugin = {
 
     this._abortExecution = false;
     this._columnsBeforeUpdateDate = 6; // So, this is the 7th column in a 1-based table column array
-    this._columnsBeforeTitle = 1;
+    // this._columnsBeforeTitle = 1;
     this._dateFormat = null;
     this._forceReprocess = false;
-    this._lastRequestTime = null;
     this._noteContents = {};
-    this._requestsCount = 0;
     this._app = app;
     // When doing mass updates, it's preferable to work with the string locally and replace the actual note content
     // less often, since each note content replace triggers a redraw of the note table & all its per-row images.
